size_t

+++ src/engraving/dom/score.cpp
@@ -5843,7 +5843,7 @@ const std::vector<Part*>& Score::parts() const
     return m_parts;
 }
 
-int Score::visiblePartCount() const
+size_t Score::visiblePartCount() const
 {
     int count = 0;
     for (const Part* part : m_parts) {
@@ -5854,7 +5854,7 @@ int Score::visiblePartCount() const
     return count;
 }
 
-int Score::visibleStavesCount() const
+size_t Score::visibleStavesCount() const
 {
     int count = 0;
     for (const Staff* staff : m_staves) {
